update
 update contacts controllers

+++ controllers/contacts.js
@@ -25,14 +25,24 @@ const favoriteSchema = Joi.object({
 
 // !=================== controllers ================
 
-const getAll = async (req, res, next) => {
-  const result = await Contact.find();
+const getAll = async (req, res) => {
+  const { _id: owner } = req.user;
+  const { page = 1, limit = 10 } = req.query;
+  const skip = (page - 1) * limit;
+
+  console.log(owner);
+
+  const result = await Contact.find({ owner }, "-createdAt -updateAT", {
+    skip,
+    limit,
+  }).populate("owner", "name email");
   res.json(result);
 };
 
-const getContactById = async (req, res, next) => {
+const getContactById = async (req, res) => {
+  const { _id: owner } = req.user;
   const { contactId } = req.params;
-  const result = await Contact.findById(contactId);
+  const result = await Contact.findById({ _id: contactId, owner });
 
   if (!result) {
     throw HttpError(404, "Not found");
@@ -40,25 +50,33 @@ const getContactById = async (req, res, next) => {
   res.json(result);
 };
 
-const create = async (req, res, next) => {
+const create = async (req, res) => {
+  const { _id: owner } = req.user;
+
   const { error } = addSchema.validate(req.body);
   if (error) {
     throw HttpError(404, "missing required name field");
   }
 
-  const result = await Contact.create(req.body);
+  const result = await Contact.create({ ...req.body, owner });
   res.status(201).json(result);
 };
 
-const updateContact = async (req, res, next) => {
+const updateContact = async (req, res) => {
+  const { _id: owner } = req.user;
+
   const { error } = updateSchema.validate(req.body);
   if (error) {
     throw HttpError(404, "missing required name field");
   }
   const { contactId } = req.params;
-  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
-    new: true,
-  });
+  const result = await Contact.findByIdAndUpdate(
+    { _id: contactId, owner },
+    req.body,
+    {
+      new: true,
+    }
+  );
 
   if (!result) {
     throw HttpError(404, "Not found");
@@ -66,15 +84,21 @@ const updateContact = async (req, res, next) => {
   res.json(result);
 };
 
-const updateFavorite = async (req, res, next) => {
+const updateFavorite = async (req, res) => {
+  const { _id: owner } = req.user;
+
   const { error } = favoriteSchema.validate(req.body);
   if (error) {
     throw HttpError(400, "missing field favorite");
   }
   const { contactId } = req.params;
-  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
-    new: true,
-  });
+  const result = await Contact.findByIdAndUpdate(
+    { _id: contactId, owner },
+    req.body,
+    {
+      new: true,
+    }
+  );
 
   if (!result) {
     throw HttpError(404, "Not found");
@@ -82,9 +106,10 @@ const updateFavorite = async (req, res, next) => {
   res.json(result);
 };
 
-const remove = async (req, res, next) => {
+const remove = async (req, res) => {
+  const { _id: owner } = req.user;
   const { contactId } = req.params;
-  const result = await Contact.findByIdAndRemove(contactId);
+  const result = await Contact.findByIdAndRemove({ _id: contactId, owner });
 
   if (!result) {
     throw HttpError(404, "Not found");
